fix: generate page.h at script load time, not pre-link
build_page.py used AddPreAction("buildprog", ...) which defers
generation until just before linking — after all .cpp files are
compiled. handlers.cpp includes page.h at compile time, so it
was always missing and the build failed with 'compilation terminated'.

Fix: call build_page() directly when the pre-build script is loaded,
which is before any source compilation begins.

Also cast millis() to (unsigned long) in LOG macros to silence
the -Wformat=2 warning on Xtensa GCC where uint32_t is 'unsigned int'
not 'unsigned long', causing a type mismatch with %lu.

+++ src/globals.h
@@ -71,9 +71,9 @@ extern TaskHandle_t ledTaskHandle;   // set in setup(); used for stack watermark
 // ---- Structured logging ----------------------------------------------------
 // All log output includes a millisecond timestamp so serial captures are
 // trivially grep-able for [ERROR] / [WARN] even without an RTC.
-#define LOG_ERROR(fmt, ...) Serial.printf("[ERROR] %7lums " fmt "\n", millis(), ##__VA_ARGS__)
-#define LOG_WARN(fmt, ...)  Serial.printf("[WARN]  %7lums " fmt "\n", millis(), ##__VA_ARGS__)
-#define LOG_INFO(fmt, ...)  Serial.printf("[INFO]  %7lums " fmt "\n", millis(), ##__VA_ARGS__)
+#define LOG_ERROR(fmt, ...) Serial.printf("[ERROR] %7lums " fmt "\n", (unsigned long)millis(), ##__VA_ARGS__)
+#define LOG_WARN(fmt, ...)  Serial.printf("[WARN]  %7lums " fmt "\n", (unsigned long)millis(), ##__VA_ARGS__)
+#define LOG_INFO(fmt, ...)  Serial.printf("[INFO]  %7lums " fmt "\n", (unsigned long)millis(), ##__VA_ARGS__)
 
 // ---- Function declarations (implemented across modules) --------------------
 void buildHeatPalette();
